Prevent BasicAuth when username/api_token are unset
Fix issue where "Authorization" header is set, even if "username" and
"api_token" are not set by the user.

Observed case where if username and api_token are not set,
`HTTPBasicAuth` is still invoked, setting header - `Authorization: Basic Og==` ("Og==" is base64 encoding of ":")

So only invoke `HTTPBasicAuth` when username AND api_token are set.